Remove sync graphs from mobile services dashboards
We are going to need them in the future again. But, instead of commenting them out, I decided to remove them because of our workflow of developing the dashboards. We do design the dashboards on Grafana and get the JSON export and put it here. Comments would make comparisons slightly harder on PRs.